feat: カテゴリバッジをDRY化し、LINE全画面に表示 (#320)
CATEGORY_CONFIGをmessage-cardから@/lib/constantsに移動し、
共有CategoryBadgeコンポーネントを作成。Google Chat・LINE両方の
カード/一覧/詳細画面でカテゴリラベルを統一表示。
