+++ apps/web/src/app/(protected)/chat-messages/[id]/page.tsx
@@ -7,8 +7,8 @@ import { ThreadView } from "@/components/chat/thread-view";
 import { ReclassifyForm } from "@/components/reclassify-form";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { getChatMessage, getChatMessages } from "@/lib/api";
+import { CATEGORY_CONFIG } from "@/lib/constants";
 import { buildMessageSearchUrl, formatDateTimeJST } from "@/lib/utils";
-import { CATEGORY_CONFIG } from "../message-card";
 import { AiPanel } from "./ai-panel";
 import { ChatOpenButton } from "./chat-open-button";
 import { ResponseStatusControl } from "./response-status-control";
