+++ apps/web/src/app/(protected)/chat-messages/message-card.tsx
@@ -2,63 +2,12 @@
 
 import { CornerDownRight, ExternalLink, Paperclip, Pencil } from "lucide-react";
 import { useRouter } from "next/navigation";
+import { CategoryBadge } from "@/components/category-badge";
 import { AttachmentList } from "@/components/chat/attachment-list";
 import { ContentWithMentions } from "@/components/chat/rich-content";
+import { CATEGORY_CONFIG } from "@/lib/constants";
 import type { ChatMessageSummary } from "@/lib/types";
 import { buildMessageSearchUrl, formatDateTimeJST } from "@/lib/utils";
-
-export const CATEGORY_CONFIG: Record<string, { label: string; accent: string; pill: string }> = {
-  salary: {
-    label: "給与・社保",
-    accent: "border-l-emerald-500",
-    pill: "bg-emerald-50 text-emerald-700 ring-1 ring-inset ring-emerald-200",
-  },
-  retirement: {
-    label: "退職・休職",
-    accent: "border-l-red-500",
-    pill: "bg-red-50 text-red-700 ring-1 ring-inset ring-red-200",
-  },
-  hiring: {
-    label: "入社・採用",
-    accent: "border-l-blue-500",
-    pill: "bg-blue-50 text-blue-700 ring-1 ring-inset ring-blue-200",
-  },
-  contract: {
-    label: "契約変更",
-    accent: "border-l-amber-500",
-    pill: "bg-amber-50 text-amber-700 ring-1 ring-inset ring-amber-200",
-  },
-  transfer: {
-    label: "施設・異動",
-    accent: "border-l-purple-500",
-    pill: "bg-purple-50 text-purple-700 ring-1 ring-inset ring-purple-200",
-  },
-  foreigner: {
-    label: "外国人・ビザ",
-    accent: "border-l-orange-500",
-    pill: "bg-orange-50 text-orange-700 ring-1 ring-inset ring-orange-200",
-  },
-  training: {
-    label: "研修・監査",
-    accent: "border-l-indigo-500",
-    pill: "bg-indigo-50 text-indigo-700 ring-1 ring-inset ring-indigo-200",
-  },
-  health_check: {
-    label: "健康診断",
-    accent: "border-l-pink-500",
-    pill: "bg-pink-50 text-pink-700 ring-1 ring-inset ring-pink-200",
-  },
-  attendance: {
-    label: "勤怠・休暇",
-    accent: "border-l-teal-500",
-    pill: "bg-teal-50 text-teal-700 ring-1 ring-inset ring-teal-200",
-  },
-  other: {
-    label: "その他",
-    accent: "border-l-slate-300",
-    pill: "bg-slate-100 text-slate-600 ring-1 ring-inset ring-slate-200",
-  },
-};
 
 const STATUS_CONFIG: Record<string, { label: string; cls: string }> = {
   unresponded: { label: "未対応", cls: "bg-rose-50 text-rose-700 ring-1 ring-inset ring-rose-200" },
@@ -212,15 +161,7 @@ export function MessageCard({ msg }: { msg: ChatMessageSummary }) {
 
             {/* Metadata */}
             <div className="flex flex-wrap items-center gap-2">
-              {catCfg ? (
-                <span
-                  className={`inline-flex items-center rounded-md px-2 py-0.5 text-xs font-medium ${catCfg.pill}`}
-                >
-                  {catCfg.label}
-                </span>
-              ) : (
-                <span className="text-xs text-muted-foreground">未分類</span>
-              )}
+              <CategoryBadge category={msg.intent?.category} />
               {methCfg && (
                 <span
                   className={`inline-flex items-center rounded-md px-2 py-0.5 text-xs font-medium ${methCfg.cls}`}
